requirements.txt
cffi==1.14.2
cryptography==3.1
pycparser==2.20
pycryptodome==3.9.8
six==1.15.0
